move select_methods to mode.py
